refactor: relocate RdfjsQuadStore implementation and tests to dedicated rdfjs provider container
