+++ src/client/providers/rdfjs/mod.ts
@@ -1,1 +1,2 @@
 export * from "./rdfjs-search-index.ts";
+export * from "./rdfjs-quad-store.ts";
