+++ src/client/quad-store/mod.ts
@@ -1,4 +1,3 @@
 export * from "./quad-store-interface.ts";
-export * from "./rdfjs-quad-store.ts";
 export * from "./patch-queue.ts";
 export * from "./patch-queue-interface.ts";
